Sketch blanket impl
