+++ tuxedo-template-runtime/src/amoeba.rs
@@ -129,31 +129,6 @@ impl SimpleVerifier for AmoebaMitosis {
     }
 }
 
-impl Verifier for AmoebaMitosis {
-    type Error = VerifierError;
-
-    fn verify<R: Redeemer>(
-        &self,
-        inputs: &[Input],
-        outputs: &[Output<R>],
-    ) -> Result<TransactionPriority, Self::Error> {
-        let input_data: Vec<DynamicallyTypedData> = inputs
-            .iter()
-            .map(|i| {
-                TransparentUtxoSet::<R>::peek_utxo(&i.output_ref)
-                    .expect("We just checked that all inputs were present.")
-                    .payload
-            })
-            .collect();
-        let output_data: Vec<DynamicallyTypedData> = outputs
-            .iter()
-            .map(|o| o.payload.clone())
-            .collect();
-
-        <AmoebaMitosis as SimpleVerifier>::verify(self, &input_data, &output_data)
-    }
-}
-
 /// A verifier for simple death of an amoeba.
 ///
 /// Any amoeba can be killed by providing it as the sole input to this verifier. No
@@ -190,31 +165,6 @@ impl SimpleVerifier for AmoebaDeath {
     }
 }
 
-impl Verifier for AmoebaDeath {
-    type Error = VerifierError;
-
-    fn verify<R: Redeemer>(
-        &self,
-        inputs: &[Input],
-        outputs: &[Output<R>],
-    ) -> Result<TransactionPriority, Self::Error> {
-        let input_data: Vec<DynamicallyTypedData> = inputs
-            .iter()
-            .map(|i| {
-                TransparentUtxoSet::<R>::peek_utxo(&i.output_ref)
-                    .expect("We just checked that all inputs were present.")
-                    .payload
-            })
-            .collect();
-        let output_data: Vec<DynamicallyTypedData> = outputs
-            .iter()
-            .map(|o| o.payload.clone())
-            .collect();
-
-        <AmoebaDeath as SimpleVerifier>::verify(self, &input_data, &output_data)
-    }
-}
-
 /// A verifier for simple creation of an amoeba.
 ///
 /// A new amoeba can be created by providing it as the sole output to this verifier. No
@@ -248,31 +198,6 @@ impl SimpleVerifier for AmoebaCreation {
         ensure!(input_data.is_empty(), VerifierError::CreationMayNotConsume);
 
         Ok(0)
-    }
-}
-
-impl Verifier for AmoebaCreation {
-    type Error = VerifierError;
-
-    fn verify<R: Redeemer>(
-        &self,
-        inputs: &[Input],
-        outputs: &[Output<R>],
-    ) -> Result<TransactionPriority, Self::Error> {
-        let input_data: Vec<DynamicallyTypedData> = inputs
-            .iter()
-            .map(|i| {
-                TransparentUtxoSet::<R>::peek_utxo(&i.output_ref)
-                    .expect("We just checked that all inputs were present.")
-                    .payload
-            })
-            .collect();
-        let output_data: Vec<DynamicallyTypedData> = outputs
-            .iter()
-            .map(|o| o.payload.clone())
-            .collect();
-
-        <AmoebaCreation as SimpleVerifier>::verify(self, &input_data, &output_data)
     }
 }
