+++ tuxedo-template-runtime/src/kitties.rs
@@ -380,32 +380,6 @@ impl SimpleVerifier for FreeKittyVerifier {
     }
 }
 
-// 5.) Eventually Abstract this with derive macro
-impl Verifier for FreeKittyVerifier {
-    type Error = VerifierError;
-
-    fn verify<R: Redeemer>(
-        &self,
-        inputs: &[Input],
-        outputs: &[Output<R>],
-    ) -> Result<TransactionPriority, Self::Error> {
-        let input_data: Vec<DynamicallyTypedData> = inputs
-            .iter()
-            .map(|i| {
-                TransparentUtxoSet::<R>::peek_utxo(&i.output_ref)
-                    .expect("We just checked that all inputs were present.")
-                    .payload
-            })
-            .collect();
-        let output_data: Vec<DynamicallyTypedData> = outputs
-            .iter()
-            .map(|o| o.payload.clone())
-            .collect();
-
-        <FreeKittyVerifier as SimpleVerifier>::verify(self, &input_data, &output_data)
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
